services/api/requirements.txt
fastapi>=0.111.0
openai>=1.30.0
uvicorn[standard]>=0.29.0
psycopg2-binary>=2.9.9
clickhouse-connect>=0.7.19
python-dotenv>=1.0.0
slowapi>=0.1.9
prometheus-fastapi-instrumentator>=6.1.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.1.0
email-validator>=2.1.0

# Testing
pytest>=8.0.0
httpx>=0.27.0
pytest-mock>=3.12.0
